requirements.txt
azure-functions
pandas
scikit-surprise
joblib